Moved the Exercise class and created the Lesson class
